feat: change long type to int in endpoints/model
Closes: FHIQC-49

+++ src/main/java/org/folio/holdingsiq/model/PackageId.java
@@ -1,6 +1,6 @@
 package org.folio.holdingsiq.model;
 
 public record PackageId(
-  long providerIdPart,
-  long packageIdPart
+  int providerIdPart,
+  int packageIdPart
 ) { }
